Generates BEAST2 xml for MultiTypeTree

Generates a .xml file to run BEAST2 using MultiTypeTree with a fixed dated phylogeny.

Usage

MultiTypeTree_xml(
  strphylo,
  n_deme,
  coal_rate,
  bit_mig_mat,
  N = 1e+07,
  thin = 1000,
  migration_history_thin = thin,
  con = stdout(),
  run_name = "$(filebase)",
  priors = "default",
  proposal_rates = c(1, 1, 1),
  cr_shape = NULL,
  cr_rate = NULL,
  mm_shape = NULL,
  mm_rate = NULL
)

Arguments

strphylo

structured phylo object giving initialisation condition for fixed tree run with MultiType Tree

coal_rate

initial estimate for coalescent rates

N

total number of MCMC iterations (including burn-in)

thin

thinning increment for the MCMC

migration_history_thin

Thinning rate for migration history samples to be saved

con

A connection object or a character string giving the location for the output xml file (stdout() prints to console)

run_name

Name for logger files to be saved as

priors

either 'default' in which lognormal priors are used or 'gamma' in which gamma/inverse gamma priors are used (and prior parameters must be specified)

proposal_rates

Relative weights of proposal operators (migration rates update : effective population size updates : migration history updates )

bit_mig_rate

initial estimate for backward-in-time migration matrix

Value

output file or file content on screen