MultiTypeTree MCMC

Runs an MCMC chain using the NodeRetype move from MultiTypeTree (Vaughan et al., 2014) using Gamma-distributed prior distributions

Usage

Vaughan_2015_MultiTypeTree_mcmc(
  N,
  strphylo,
  coal_rate,
  bit_mig_mat,
  cr_mode,
  cr_var,
  mm_mode,
  mm_var,
  st_depth = 1,
  output_dir,
  run_name = "MultiTypeTree",
  stdout_log = TRUE,
  thin = max(N/5000, 1),
  save_migration_history = TRUE,
  migration_history_thin = thin,
  proposal_rates = c(1000, 1, 1)
)

Arguments

N

Total number of MCMC iterations to complete

strphylo

Initial structured phylogenetic tree

coal_rate

Initial estimate of coalescent rates

bit_mig_mat

Initial estimate of backward-in-time migration rates matrix

cr_mode

Coalescent rates prior mode

cr_var

Coalescent rates prior variance

mm_mode

Backward-in-time migration rates prior mode

mm_var

Backward-in-time migration rates prior variance

output_dir

Directory to output log files to

run_name

Run name to save log files as

stdout_log

(logical) Output logs to stdout()? Default true

thin

Thinning rate for continuous parameter posterior samples

save_migration_history

Logical value indicating whether tree samples are saved (i.e. whether .trees file is created)

migration_history_thin

Thinning rate for tree samples to be saved

proposal_rates

Relative rates of migration history and continuous parameter updates (migration history : coalescent rates : migration rates)