The Truncated Exponential Distribution

Density, distribution function and random generation for the exponential distribution with rate rate truncated above by c

Usage

dexp_trunc(x, c = Inf, rate = 1)

pexp_trunc(q, c = Inf, rate = 1)

rexp_trunc(n, c = Inf, rate = 1)

Arguments

x, q

vector of quantiles

c

truncation cutoff

rate

vector of rates

n

number of observations